불필요한 기능 제거 및 일부 수정!

+++ src/main/java/com/forweaver/util/WebUtil.java
@@ -268,8 +268,8 @@ public class WebUtil {
 		if (sec < 60) {
 			min = 0;
 		} else {
-			min = (int) (sec / 60);
-			sec = (int) (sec % 60);
+			min = sec / 60;
+			sec = sec % 60;
 		}
 
 		if (min < 60) {
